requirements.txt
Flask>=2.2
Flask-WTF>=1.1.1
WTForms>=3.0.1
Flask-Login>=0.6.2
Flask-Bcrypt>=1.0.1
Flask-Migrate>=4.0.4
Flask-SQLAlchemy>=3.0.3
python-dotenv>=1.0.0
requests>=2.31.0
